refactor(runner.gui): separate code into packages
Separate Selenium and FBA code into separate packages.
