+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/gui/application/fluxbalanceanalysis/FBASeleniumInputProgrammer.java
@@ -1,7 +1,11 @@
-package edu.iastate.ece.sd.sdmay2126.runner.gui;
+package edu.iastate.ece.sd.sdmay2126.runner.gui.application.fluxbalanceanalysis;
 
 import edu.iastate.ece.sd.sdmay2126.application.FBAParameters;
 import edu.iastate.ece.sd.sdmay2126.orchestration.Job;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.KBaseHelper;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.selenium.SeleniumIdentificationException;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.selenium.SeleniumInputProgrammer;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.selenium.SeleniumUtilities;
 import org.openqa.selenium.By;
 import org.openqa.selenium.TimeoutException;
 import org.openqa.selenium.WebDriver;
