+++ src/main/java/edu/iastate/ece/sd/sdmay2126/runner/gui/authentication/globus/GlobusAuthenticationFlow.java
@@ -1,7 +1,7 @@
 package edu.iastate.ece.sd.sdmay2126.runner.gui.authentication.globus;
 
-import edu.iastate.ece.sd.sdmay2126.runner.gui.SeleniumIdentificationException;
-import edu.iastate.ece.sd.sdmay2126.runner.gui.SeleniumUtilities;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.selenium.SeleniumIdentificationException;
+import edu.iastate.ece.sd.sdmay2126.runner.gui.selenium.SeleniumUtilities;
 import edu.iastate.ece.sd.sdmay2126.runner.gui.authentication.SeleniumAuthenticationFlow;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
